refactor: move all NVM3 stuff into one directory
