+++ packages/nvmedit/src/index.ts
@@ -17,13 +17,8 @@ export type {
 	NVMJSONNodeWithInfo,
 	NVMJSONVirtualNode,
 } from "./convert";
-export {
-	ControllerInfoFile,
-	ControllerInfoFileID,
-} from "./files/ControllerInfoFile";
-export { NVMFile } from "./files/NVMFile";
 export { NVM3 } from "./lib/NVM3";
-export { NVM3Adapter } from "./lib/NVM3Adapter";
+export type { NVM3Meta as NVMMeta } from "./lib/NVM3";
 export type {
 	ControllerNVMProperty,
 	ControllerNVMPropertyToDataType,
@@ -40,18 +35,23 @@ export type {
 export { NVMAccess } from "./lib/common/definitions";
 export { BufferedNVMReader } from "./lib/io/BufferedNVMReader";
 export { NVMFileIO } from "./lib/io/NVMFileIO";
+export { NVM3Adapter } from "./lib/nvm3/adapter";
 export {
 	FragmentType,
 	ObjectType,
 	PageStatus,
 	PageWriteSize,
-} from "./nvm3/consts";
-export type { NVMMeta } from "./nvm3/nvm";
-export type { NVM3Object as NVMObject } from "./nvm3/object";
+} from "./lib/nvm3/consts";
+export {
+	ControllerInfoFile,
+	ControllerInfoFileID,
+	NVMFile,
+} from "./lib/nvm3/files";
+export type { NVM3Object as NVMObject } from "./lib/nvm3/object";
 export type {
 	NVM3Page as NVMPage,
 	NVM3PageHeader as PageHeader,
-} from "./nvm3/page";
+} from "./lib/nvm3/page";
 export type {
 	NVM500JSON,
 	NVM500JSONController,
