+++ packages/nvmedit/src/lib/nvm3/consts.ts
