+++ packages/nvmedit/src/lib/nvm3/files/ApplicationDataFile.ts
