+++ packages/nvmedit/src/lib/nvm3/files/ProtocolNodeMaskFiles.ts
@@ -1,5 +1,5 @@
 import { NODE_ID_MAX, encodeBitMask, parseBitMask } from "@zwave-js/core/safe";
-import type { NVM3Object } from "../nvm3/object";
+import type { NVM3Object } from "../object";
 import {
 	NVMFile,
 	type NVMFileCreationOptions,
