+++ packages/nvmedit/src/lib/nvm3/files/index.ts
@@ -2,6 +2,7 @@ import "reflect-metadata";
 
 export * from "./ApplicationCCsFile";
 export * from "./ApplicationDataFile";
+export * from "./ApplicationNameFile";
 export * from "./ApplicationRFConfigFile";
 export * from "./ApplicationTypeFile";
 export * from "./ControllerInfoFile";
